backend/requirements.txt
# backend/requirements.txt
# Alone AI — backend kutubxonalari

# ── FastAPI ───────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────
# aiosqlite olib tashlandi — sync sqlite ishlatilmoqda

# ── Auth & Security ───────────────────────────────────────────
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
# python-jose olib tashlandi — PyJWT yetarli

# ── AI Modellar ───────────────────────────────────────────────
groq==0.9.0
google-generativeai==0.7.2
openai==1.35.0

# ── TTS/STT ───────────────────────────────────────────────────
edge-tts==6.1.9

# ── HTTP ──────────────────────────────────────────────────────
httpx==0.27.0

# ── Fayllar ───────────────────────────────────────────────────
pypdf==4.2.0
python-docx==1.1.2
# python-magic olib tashlandi — Windows da muammo
# Windows da kerak bo'lsa: pip install python-magic-bin

# ── 2FA ───────────────────────────────────────────────────────
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.3.0

# ── To'lov ────────────────────────────────────────────────────
stripe==9.12.0

# ── Email ─────────────────────────────────────────────────────
aiosmtplib==3.0.1

# ── Qidiruv ───────────────────────────────────────────────────
tavily-python==0.3.3

# ── Monitoring ────────────────────────────────────────────────
psutil==5.9.8

# ── Utils ─────────────────────────────────────────────────────
pydantic[email]==2.7.1
pydantic-settings==2.3.1
python-dateutil==2.9.0
python-dotenv==1.0.1

# ── Dev/Test ──────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7